Bolt and Repulsor
Added Bolt (fully functional) and made Repulsor fully work.

Repulsor can be used by right clicking with a carrot on a stick in your offhand while holding a heart of the sea in your mainhand.
Bolt can be used by right clicking with a carrot on a stick in your offhand while holding a diamond in your mainhand.

Repulsor creates a shield that deflects arrows back at their owner, or are simply killed if no owner is found. Intended to also push away players who are in the radius but delta API isn't added yet so it currently doesn't do that.
Bolt causes glow and some FX on use, then 15 ticks later fires a fast-moving lightning bolt in a straight line in front of you. Behaves much like its counterpart in Rogues 2 but with some slight tweaks to fit the game better.